stock_predictor_app/backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8.0
scikit-learn>=1.2.0
joblib>=1.2.0
python-multipart>=0.0.6 